Fixing bugs after merge, refactoring
