+++ src/pages/packsList/PackReducer.ts
@@ -1,6 +1,6 @@
-import {createSlice, PayloadAction} from "@reduxjs/toolkit";
+import {createSlice} from "@reduxjs/toolkit";
 import {
-    APIPacks, CardPacksEntityWithDeckCover,
+    APIPacks,
     CreateNewCardPackPayload,
     getCardPacksPayload,
     GetCardsPackResponse,
@@ -8,8 +8,7 @@ import {
 } from "../../DAL/API-CardsPack";
 import {AppDispatchType, AppThunk} from "../app/store";
 import {HandleToggleStatusAppAndInterceptorErrors} from "../../utils/HandleToggleStatusAppAndInterceptorErrors";
-import {ExtendedCardEntity, getCardsPayload, GetCardsResponse} from "../../DAL/API-Cards";
-import {actionsCards} from "../cardsList/CardsReducer";
+import { HeadCell } from "../../common/components/table/CommonTable";
 
 
 type InitialState = {
@@ -17,18 +16,6 @@ type InitialState = {
     queryParams: getCardPacksPayload
     initHeadCells: HeadCell[]
 }
-
-
-//type Numeric = "inherit" | "right" | "left" | "center" | "justify" | undefined;
-
-export interface HeadCell {
-    numeric: "inherit" | "right" | "left" | "center" | "justify" | undefined;
-    //id: keyof CardPacksEntityWithDeckCover|"action";
-    id: string,
-    label: string;
-    order: "0" | "1" | undefined
-}
-
 
 const initialState: InitialState = {
     packsData: {} as GetCardsPackResponse,
