+++ src/pages/packsList/PacksPage.tsx
@@ -2,12 +2,12 @@ import * as React from "react";
 import {useDispatchApp, useSelectorApp} from "../../CustomHooks/CustomHooks";
 import {useEffect} from "react";
 import {thunksPack} from "./PackReducer";
-import {TablePacks1} from "./TablePacks1";
+import {TablePacks} from "./TablePacks";
 
 
 export const PacksPage = () => {
 
-    const cardsPack_id = useSelectorApp(state => state.cards.queryParams.cardsPack_id)
+
     const packs = useSelectorApp(state => state.packs.packsData.cardPacks)
     const headCells = useSelectorApp(state => state.packs.initHeadCells)
 
@@ -21,7 +21,7 @@ export const PacksPage = () => {
         dispatch(thunksPack.createPack({name: 'програмист! иди спать, ты пьян!'}))
     }
     return (
-        <TablePacks1 packs={!!packs ? packs : []} headCells={headCells}/>
+        <TablePacks packs={!!packs ? packs : []} headCells={headCells}/>
     )
 }
